VueComponents: Convert stories with *.js extension to *.ts
